Django/Practice/car_dealer/requirements/base.txt
Django==3.1.7
django-environ==0.4.5
psycopg2-binary==2.8.6
Pillow==8.1.2
djangorestframework==3.12.4